Move pragma handlers (NFC)
